pyproject.toml
[project]
name = "eventos-api"
version = "0.1.0"
description = "Backend FastAPI Eventos"
authors = [
    {name = "CBarrantes24",email = "103292667+CBarrantes24@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry]
package-mode = false

